build and distribute lambda archive (#1)
Use a pulumi program to build the lambda archive and then orchestrate signing it and distributing into regional buckets, where it can be directly referenced for customer deployments. (see #2)

The archive is built in a docker container to ensure toolchain consistency, and then exported locally.
It is then uploaded to a staging s3 bucket, (named after the githash of the lastest change), and signed.
The signed archive is then distributed into public regional buckets.

Resolves pulumi/pulumi-service#26264