rewrite lambda in go (#8)
Initially I thought it would be nice for the lambda to be written in
typescript because it would let users inspect the code in the console,
but it turns out that when codesigning is used the code is not visible
(unless you download the zip):

Writing the lambda in go will allow us to reuse the implementation for
the direct connect variant by importing it as a lib.

I've updated the build and the deploy pulumi programs to handle the
golang variant. Rerunning deploy updated my existing typescript lambda
to the go one, and it invokes successfully.

I've also included test cases that leverage
[testcontainers](https://golang.testcontainers.org/quickstart/) for
spinning up a temporary db to validate the rotation implementation
actually works.